Moved album card and made a mainHeaderCard for mainpage
