+++ client/components/layout/cards/artistCard.tsx
@@ -1,14 +1,16 @@
 import { Box, Card } from "@mui/material";
-import { Header3 } from "../textComponents";
-import DefaultButton from "../buttons/defaulButton";
+import { Header3 } from "../../textComponents";
+import DefaultButton from "../../buttons/defaulButton";
 import theme from "@/utils/theme";
-import Track from "@/types/trackTypes";
+import Artist from "@/types/artistTypes";
 
-export default function TrackCard(props: { track: Track }) {
+export default function ArtistCard(props: { artist: Artist }) {
+    const imageUrl = props.artist.icon_link
+
     return (
         <Card
             sx={{
-                backgroundColor: theme.palette.secondary.main,
+                backgroundImage: `url(${imageUrl})`,
                 position: 'absolute',
                 top: 0,
                 left: 0,
@@ -18,7 +20,7 @@ export default function TrackCard(props: { track: Track }) {
                 flexDirection: 'column',
                 justifyContent: 'space-between',
             }}>
-            <Header3 text={props.track.name} />
+            <Header3 text={props.artist.name} />
             <Box sx={{
                 marginLeft: 'auto',
                 marginRight: 1,
