+++ client/components/layout/cards/trackCard.tsx
@@ -1,16 +1,14 @@
 import { Box, Card } from "@mui/material";
-import { Header3 } from "../textComponents";
-import DefaultButton from "../buttons/defaulButton";
+import { Header3 } from "../../textComponents";
+import DefaultButton from "../../buttons/defaulButton";
 import theme from "@/utils/theme";
-import Playlist from "@/types/playlistTypes";
+import Track from "@/types/trackTypes";
 
-export default function PlaylistCard(props: { playlist: Playlist }) {
-    const imageUrl = props.playlist.icon_link
-
+export default function TrackCard(props: { track: Track }) {
     return (
         <Card
             sx={{
-                backgroundImage: `url(${imageUrl})`,
+                backgroundColor: theme.palette.secondary.main,
                 position: 'absolute',
                 top: 0,
                 left: 0,
@@ -20,7 +18,7 @@ export default function PlaylistCard(props: { playlist: Playlist }) {
                 flexDirection: 'column',
                 justifyContent: 'space-between',
             }}>
-            <Header3 text={props.playlist.name} />
+            <Header3 text={props.track.name} />
             <Box sx={{
                 marginLeft: 'auto',
                 marginRight: 1,
